Split Batcher logic with OpenGl implementation
Following decisions in go/stadia-orbit-testing-ui-elements, we want to
refactor Batcher functionality to make it easier to mock. In this PR we
are creating OpenGlBatcher (first step in http://b/225173189).

In addition we cleaned includes and refactor virtual methods to not have
default parameters.

Next step: Create PrimitiveAssembler.
