+++ src/OrbitGl/CMakeLists.txt
@@ -22,6 +22,7 @@ target_sources(
          AsyncTrack.h
          BasicPageFaultsTrack.h
          Batcher.h
+         BatcherImpl.h
          CallstackThreadBar.h
          CallTreeView.h
          CaptureStats.h
@@ -95,6 +96,7 @@ target_sources(
           AsyncTrack.cpp
           BasicPageFaultsTrack.cpp
           Batcher.cpp
+          BatcherImpl.cpp
           CallstackThreadBar.cpp
           CallTreeView.cpp
           CaptureStats.cpp
